Added Blazed.Toast, Completed Vertical Slice for ProjectCreate and ProjectList Components, Refactored out EditorComponentBase and ListComponentBase

+++ ComicMaker/Common/Models/PagedResult.cs
@@ -5,9 +5,9 @@ namespace ComicMaker.Common.Models
     public class PagedResult<T>
     {
         public IList<T> Items { get; set; } = new List<T>();
-        public long CurrentPage { get; set; }
-        public long TotalPages { get; set; }
-        public long TotalItems { get; set; }
-        public long ItemsPerPage { get; set; }
+        public int CurrentPage { get; set; }
+        public int TotalPages { get; set; }
+        public int TotalItems { get; set; }
+        public int ItemsPerPage { get; set; }
     }
 }
